# api/swagger.yaml
swagger: '2.0'
info:
  description: 南大测试 REST API
  version: 2.0.1
  title: 南大测试 REST API 接口说明
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8000'
...
schemes:
  - http
consumes:
  - application/x-www-form-urlencoded
produces:
  - application/json
paths:
...
  /services/v1/testCase:
    get:
      tags:
        - testCase
      summary: 查询测试方案
      description: >-
        根据用户查询或者根据工程查询，从传入参数里是否有projectID区分，如果没有则返回当前用户所有测试方案或者管理员获取所有测试方案，如果给了projectID则返回该工程里的测试方案
      operationId: queryTestCases
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - name: projectID
          in: query
          description: 工程id，需要按照工程查找时填入该参数
          type: string
          required: false
      responses:
        '200':
          description: |
            成功查询测试方案时返回测试方案
            查询失败时，返回一个失败的status
            ```
            {
              
              "data": [
                {
                  "createdUserId": "2",
                  "createdTime": "2018-06-02 20:16:29",
                  "id": "cdd340fa-501c-42ab-9596-6c55b1a694c2"
                }
              ],
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
    post:
      tags:
        - testCase
      summary: 添加测试方案
      description: >-
        向数据库中添加测试方案，除了内容body还要前端提供project的id，这个projectID和username、clientDigest一样的形式传入，成功时返回测试方案的具体信息以及一个success状态
      operationId: addTestCase
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - name: projectID
          in: query
          description: 绑定的工程的id
          type: string
          required: true
        - in: formData
          name: params
          description: |
            测试方案的具体内容. JSON格式
            ```
            {
              "body": "xxx"
            }
            ```
          type: string
          required: true
      responses:
        '200':
          description: |
            成功添加测试方案时的样例response body
            ```
            {
              "data": {
                "createdUserId": "2",
                "createdTime": "2018-06-03 20:16:31",
                "id": "49e9b994-064c-47f9-9a2e-f02260ed9d11",
                "body": "I love taylorswift"
              },
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
    put:
      tags:
        - testCase
      summary: 修改测试方案
      description: 修改已有的测试方案，参数中需要包含测试方案的id
      operationId: editTestCase
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - in: formData
          name: params
          description: |
            具体内容. JSON格式. 需要指明测试方案的ID
            ```
            {
              "id": 
              "body":
            }
            ```
          type: string
          required: true
      responses:
        '200':
          description: |
            修改成功时的样例response body
            ```
            {
              "data": {
                "createdUserId": "2",
                "alteredTime": "2018-06-03 20:17:52",
                "createdTime": "2018-06-03 20:16:31",
                "alteredUserId": "2",
                "id": "49e9b994-064c-47f9-9a2e-f02260ed9d11",
                "body": "love shawnmendes"
              },
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
    delete:
      tags:
        - testCase
      summary: 删除测试方案
      description: 删除一个测试方案，需要提供具体id
      operationId: deleteTestCase
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - in: formData
          name: params
          description: |
            具体内容. JSON格式. 需要指明ID,其他内容为非必要
            ```
            {
              "id": 
              "xxx":
            }
            ```
          type: string
          required: true
      responses:
        '200':
          description: |
            成功删除测试方案
            ```
            {
              "status":"SUCCESS"
            }
            ```
        '405':
          description: Invalid input
  '/services/v1/testCase/{id}':
    get:
      tags:
        - testCase
      summary: 根据id查询测试方案
      description: 根据请求中的id查询对应测试方案
      operationId: queryTestCaseById
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 测试方案的id
          type: string
          required: true
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
      responses:
        '200':
          description: |
            成功查询测试方案. data内包含JsonArray格式的内容
            ```
            {
              "data": {
                "createdUserId": "2",
                "createdTime": "2018-06-03 20:19:30",
                "id": "ad402346-a514-4e0b-9fec-6ac13f0a2a98",
                "testcase": "I love taylorswift"
              },
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
...
  /services/v1/testFunction:
    get:
      tags:
        - testFunction
      summary: 查询测试功能
      description: >-
        根据用户查询或者根据工程查询，从传入参数里是否有projectID区分，如果没有则返回当前用户所有测试功能或者管理员获取所有测试功能，如果给了projectID则返回该工程里的测试功能
      operationId: queryTestFunctions
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - name: projectID
          in: query
          description: 工程id，需要按照工程查找时填入该参数
          type: string
          required: false
      responses:
        '200':
          description: |
            成功查询测试功能时返回测试功能
            查询失败时，返回一个失败的status
            ```
            {
              
              "data": [
                {
                  "createdUserId": "2",
                  "createdTime": "2018-06-02 20:16:29",
                  "id": "cdd340fa-501c-42ab-9596-6c55b1a694c2"
                }
              ],
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
    post:
      tags:
        - testFunction
      summary: 添加测试功能
      description: >-
        添加测试方案，除了内容body还要前端提供project的id，这个projectID和username、clientDigest一样的形式传入，成功时返回测试方案的具体信息以及一个success状态
      operationId: addTestFunction
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - name: projectID
          in: query
          description: 绑定的工程的id
          type: string
          required: true
        - in: formData
          name: params
          description: |
            测试功能的具体内容. JSON格式
            ```
            {
              "body": "xxx"
            }
            ```
          type: string
          required: true
      responses:
        '200':
          description: |
            成功添加测试功能时的样例response body
            ```
            {
              "data": {
                "createdUserId": "2",
                "createdTime": "2018-06-03 20:16:31",
                "id": "b68cd9c6-fb4b-4bff-9cbe-f1804fcc6678",
                "body": "22"
              },
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
    put:
      tags:
        - testFunction
      summary: 修改测试功能
      description: 修改已有的测试功能，参数中需要包含测试功能的id
      operationId: editTestFunction
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - in: formData
          name: params
          description: |
            具体内容. JSON格式. 需要指明测试功能的ID
            ```
            {
              "id": 
              "body":
            }
            ```
          type: string
          required: true
      responses:
        '200':
          description: |
            修改成功时的样例response body
            ```
            {
              "data": {
                "createdUserId": "2",
                "alteredTime": "2018-06-03 20:17:52",
                "createdTime": "2018-06-03 20:16:31",
                "alteredUserId": "2",
                "id": "49e9b994-064c-47f9-9a2e-f02260ed9d11",
                "body": "i don't know about you"
              },
              "status": "SUCCESS"
            }
            ```
        '405':
          description: Invalid input
    delete:
      tags:
        - testFunction
      summary: 删除测试功能
      description: 删除一个测试功能，需要提供具体id
      operationId: deleteTestFunction
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          description: 用户名
          type: string
          required: true
        - name: clientDigest
          in: query
          description: 摘要。登录时由后台传入，用以身份验证
          type: string
          required: true
        - in: formData
          name: params
          description: |
            具体内容. JSON格式. 需要指明ID,其他内容为非必要
            ```
            {
              "id": 
              "xxx":
            }
            ```
          type: string
          required: true
      responses:
        '200':
          description: |
            成功删除测试功能
            ```
            {
              "status":"SUCCESS"
            }
            ```
        '405':
          description: Invalid input
...
securityDefinitions:
  petstore_auth:
    type: oauth2
    authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
    flow: implicit
    scopes:
      'write:pets': modify pets in your account
      'read:pets': read your pets
  api_key:
    type: apiKey
    name: api_key
    in: header
...